Default parallel movement punishment is in the table with punishments.

+++ CounterpointComposer/src/main/java/pl/szachewicz/algorithm/Evaluator.java
@@ -42,7 +42,7 @@ public class Evaluator {
 						&& counterpointPitch < counterpointPreviousPitch)) {
 
 					int semitones = Math.abs(cantusFirmusPitch - counterpointPitch);
-					int punishment = preferences.getPunishmentForParallelMovement(semitones);
+					double punishment = preferences.getPunishmentForParallelMovement(semitones);
 
 					points -= punishment;
 					log(i-1, i, "parallel movement", punishment);
